refactor(ats-validator): split index.js into focused modules (behavior-preserving)
