+++ packages/ats-validator/src/grade.js
@@ -1,0 +1,10 @@
+/**
+ * Get letter grade from percentage
+ */
+export function getGrade(percentage) {
+  if (percentage >= 90) return 'A';
+  if (percentage >= 80) return 'B';
+  if (percentage >= 70) return 'C';
+  if (percentage >= 60) return 'D';
+  return 'F';
+}
